+++ wrappers/src/set_cs.cpp
@@ -19,12 +19,15 @@
 #include <realm.hpp>
 #include <realm/object-store/object_accessor.hpp>
 #include <realm/object-store/thread_safe_reference.hpp>
+#include <realm/parser/query_parser.hpp>
 
 #include "error_handling.hpp"
+#include "filter.hpp"
 #include "marshalling.hpp"
 #include "realm_export_decls.hpp"
 #include "wrapper_exceptions.hpp"
 #include "notifications_cs.hpp"
+#include "schema_cs.hpp"
 
 using namespace realm;
 using namespace realm::binding;
@@ -258,6 +261,20 @@ REALM_EXPORT bool realm_set_set_equals(object_store::Set& set, const object_stor
 {
     return handle_errors(ex, [&]() -> bool {
         return set.set_equals(hack_remove_me(other_collection));
+    });
+}
+
+REALM_EXPORT Results* set_to_results(const object_store::Set& set, NativeException::Marshallable& ex)
+{
+    return handle_errors(ex, [&]() {
+        return new Results(set.as_results());
+    });
+}
+
+REALM_EXPORT Results* set_get_filtered_results(const object_store::Set& set, uint16_t* query_buf, size_t query_len, realm_value_t* arguments, size_t args_count, NativeException::Marshallable& ex)
+{
+    return handle_errors(ex, [&]() {
+        return get_filtered_results(set.get_realm(), set.get_table(), set.get_query(), query_buf, query_len, arguments, args_count, DescriptorOrdering());
     });
 }